requirements.txt
streamlit==1.34.0
torch>=2.0
torchvision>=0.15
transformers>=4.36
pillow
numpy
